app_server: fix clang warning tautological-constant-out-of-range-compare
Clang message: error: result of comparison of constant -1 with expression
of type 'direct_driver_state' is always true

Reviewed-on: https://review.haiku-os.org/c/haiku/+/2394

+++ src/servers/app/DirectWindowInfo.cpp
@@ -89,7 +89,7 @@ DirectWindowInfo::SetState(direct_buffer_state bufferState,
 
 	fBufferInfo->buffer_state = bufferState;
 
-	if (driverState != -1)
+	if ((int)driverState != -1)
 		fBufferInfo->driver_state = driverState;
 
 	if ((bufferState & B_DIRECT_MODE_MASK) != B_DIRECT_STOP) {
